svg: fixed vertex glitch (bad absolute position was pushed). And experiment smooth line by default for strokes.
